+++ src/mesido/esdl/profile_parser.py
@@ -197,6 +197,8 @@ class InfluxDBProfileReader(BaseProfileReader):
         esdl.esdl.HeatingDemand: ".target_heat_demand",
         esdl.esdl.GenericConsumer: ".target_heat_demand",
         esdl.esdl.HeatProducer: ".maximum_heat_source",
+        esdl.esdl.GeothermalSource: ".maximum_heat_source",
+        esdl.esdl.ResidualHeatSource: ".maximum_heat_source",
         esdl.esdl.ElectricityDemand: ".target_electricity_demand",
         esdl.esdl.ElectricityProducer: ".maximum_electricity_source",
         esdl.esdl.GasDemand: ".target_gas_demand",
@@ -290,7 +292,11 @@ class InfluxDBProfileReader(BaseProfileReader):
             )
 
             container = profile.eContainer()
-            if isinstance(container, esdl.Commodity):
+            if isinstance(container, esdl.ProfileConstraint):
+                asset = container.eContainer()
+                variable_suffix = self.asset_type_to_variable_name_conversion[type(asset)]
+                var_base_name = asset.name
+            elif isinstance(container, esdl.Commodity):
                 variable_suffix = self.carrier_profile_var_name
                 var_base_name = container.name
             elif isinstance(container, esdl.Port):